♻️ 30TKV4 cli: split preview and intake test responsibilities
