+++ .agentplane/tasks/202608061646-30TKV4/pr/diffstat.txt
@@ -2,16 +2,18 @@
  packages/agentplane/src/cli/command-invocations.ts |   1 +
  .../src/cli/run-cli.core.route-decision.test.ts    |   8 +
  .../src/cli/run-cli.core.task-run.test.ts          | 112 +++---
- .../src/cli/run-cli.core.tasks.create.test.ts      | 282 ++++++++++++++
+ .../src/cli/run-cli.core.tasks.create.test.ts      |   5 -
+ .../src/cli/run-cli.core.tasks.user-create.test.ts | 302 +++++++++++++++
  ...-cli.critical.agent-efficiency-baseline.test.ts |  39 +-
  .../src/cli/run-cli/command-catalog/task.ts        |   9 +
  .../src/cli/run-cli/command-loaders/task.ts        |   4 +
  .../agentplane/src/commands/task/create.command.ts | 409 +++++++++++++++++++++
  packages/agentplane/src/commands/task/new.ts       | 333 +++++++++--------
- .../agentplane/src/commands/task/run-render.ts     | 122 +++++-
+ .../src/commands/task/run-execution-preview.ts     |  83 +++++
+ .../agentplane/src/commands/task/run-render.ts     |  39 +-
  .../agentplane/src/commands/task/run.command.ts    |   6 +-
  .../agentplane/src/commands/task/status.command.ts |  13 +-
  .../agentplane/src/commands/task/task.command.ts   |  10 +-
  .../baselines/v0.7-compatibility-candidate.json    | 193 +++++++++-
  .../check-compatibility-contract-baseline.mjs      | 135 ++++++-
- 16 files changed, 1470 insertions(+), 293 deletions(-)
+ 18 files changed, 1491 insertions(+), 297 deletions(-)
