+++ .agentplane/tasks/202608061646-30TKV4/pr/meta.json
@@ -2,7 +2,7 @@
   "base": "main",
   "branch": "task/202608061646-30TKV4/add-user-first-task-intake-and-execution-preview",
   "created_at": "2026-08-06T19:12:24.484Z",
-  "diffstat_sha256": "sha256:cc351a6b923988f98872ed6572ae06debf4c3adba8841d748d15d408f89ddf38",
+  "diffstat_sha256": "sha256:1602b8b93a0ef9a97d7e4ae369c478b22c4ccbcd248542af07c8a79a572eaac6",
   "last_verified_at": "2026-08-07T00:33:33.660Z",
   "last_verified_diffstat_sha256": "sha256:cc351a6b923988f98872ed6572ae06debf4c3adba8841d748d15d408f89ddf38",
   "pr_number": 4787,
