+++ packages/agentplane/src/commands/task/run.command.ts
@@ -28,13 +28,13 @@ import {
   renderRunnerDiagnosticInspectPayload,
   renderRunnerDiagnosticStatusPayload,
   renderRunnerStatusPayload,
-  buildTaskRunExecutionPreview,
   runnerReconciliationWarning,
   renderTaskRunnerLifecyclePayload,
   reportExecutedTaskRun,
   reportPreparedTaskRun,
   tailText,
 } from "./run-render.js";
+import { buildTaskRunExecutionPreview } from "./run-execution-preview.js";
 import { followRunnerLogs } from "./run-logs-follow.js";
 import { superviseBranchTaskRun } from "./branch-task-supervisor.js";
 import { superviseDirectTaskRun } from "./direct-task-supervisor.js";
